Last commits cleanup

+++ src/main/java/net/Indyuce/mmocore/api/droptable/dropitem/DropTableDropItem.java
@@ -4,6 +4,7 @@ import org.apache.commons.lang.Validate;
 
 import net.Indyuce.mmocore.MMOCore;
 import net.Indyuce.mmocore.api.droptable.DropTable;
+import net.Indyuce.mmocore.api.droptable.condition.ConditionInstance;
 import net.Indyuce.mmocore.api.loot.LootBuilder;
 import net.mmogroup.mmolib.api.MMOLineConfig;
 
@@ -22,7 +23,8 @@ public class DropTableDropItem extends DropItem {
 
 	@Override
 	public void collect(LootBuilder builder) {
-		for (int j = 0; j < rollAmount(); j++)
-			builder.addLoot(dropTable.collect(builder));
+		if (dropTable.areConditionsMet(new ConditionInstance(builder.getEntity().getPlayer())))
+			for (int j = 0; j < rollAmount(); j++)
+				builder.addLoot(dropTable.collect(builder));
 	}
 }
